show number of contributors instead of number of (active) tellers
A contributor is a teller who wrote up until now at least
one part of the story. This means that:
- If a user leaves a story, he/she may still be counted as contributor.
- If a user leaves a story without contributing anything,
he/she is not counted.

For finished stories, now the number of contributors is shown instead of
the number of users which have not left the story at its end.

This fixes #18.